Finish implementing FourBitRAM. Initial Test Pass!
I combed through the FourBitRAM several times, and finally got the bugs
worked out. Several of the logic gates in the class were not initialized
in the constructor, and also there was a bug with the Write Enable
Inverter.

Sometimes, because of the way things are over there at
Texas-Instruments, certain inputs come active low, and have an inverter
right before any logic input. I've had a tough time making sure that I'm
not unintentially inverting some inputs, and it seems everything is
working! I'll run through more test cases, to see if we get any
unexpected results. This is cool!